requirements.txt
attrs==22.1.0
aws-cdk-lib==2.53.0
aws-cdk.asset-awscli-v1==2.2.25
aws-cdk.asset-kubectl-v20==2.1.1
aws-cdk.asset-node-proxy-agent-v5==2.0.31
boto3==1.26.21
botocore==1.29.21
cattrs==22.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
constructs==10.1.177
exceptiongroup==1.0.4
idna==3.4
iniconfig==1.1.1
jmespath==1.0.1
jsii==1.71.0
packaging==21.3
pluggy==1.0.0
publication==0.0.3
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.13
